Switch over to new Color struct as default.
Former-commit-id: 1cb78f9723ddcdd4bcf45c27d7571423629ad0c6
Former-commit-id: 727c1b5eeaec657fa541ef64e118006c2b74911f
Former-commit-id: 7c9b65f6d71f10e9c3ce8193ae797f30b8c291f2

+++ src/ImageProcessor/Formats/Gif/Quantizer/QuantizedImage.cs
@@ -69,7 +69,7 @@ namespace ImageProcessor.Formats
             Image image = new Image();
 
             int pixelCount = this.Pixels.Length;
-            byte[] bgraPixels = new byte[pixelCount * 4];
+            float[] bgraPixels = new float[pixelCount * 4];
 
             for (int i = 0; i < pixelCount; i++)
             {
